Fixed overflow by converting int_rep into python int

+++ phe/encoding.py
@@ -188,8 +188,8 @@ class EncodedNumber(object):
             exponent = min(max_exponent, prec_exponent)
 
         # Use rationals instead of floats to avoid overflow.
-        int_rep = round(fractions.Fraction(scalar)
-                        * fractions.Fraction(cls.BASE) ** -exponent)
+        int_rep = int(round(fractions.Fraction(scalar)
+                        * fractions.Fraction(cls.BASE) ** -exponent))
 
         if abs(int_rep) > public_key.max_int:
             raise ValueError('Integer needs to be within +/- %d but got %d'
